Delete old NL interface [Part 1] (#4180)
This has not been used in a long time and deleting this will simplify
things. This PR is for frontend, backend cleanup in next part.

Testing:

![image](https://github.com/datacommonsorg/website/assets/4375037/56d180b4-3707-42bd-a820-e1f11da03eae)
And a test query on explore page still LG.

NOTE: I did not want to mess with css, in the fear of breaking
something. Left a TODO.
