+++ server/routes/nl/html.py
@@ -25,29 +25,6 @@ from flask import render_template
 bp = Blueprint('nl', __name__, url_prefix='/nl')
 
 
-@bp.route('/')
-def page():
-  placeholder_query = ''
-  # TODO: Make this more customizable for all custom DC's
-  if g.env == 'climate_trace':
-    placeholder_query = 'Greenhouse gas emissions in USA'
-  return render_template('/nl_interface.html',
-                         maps_api_key=current_app.config['MAPS_API_KEY'],
-                         placeholder_query=placeholder_query,
-                         index_type="",
-                         website_hash=os.environ.get("WEBSITE_HASH"))
-
-
-@bp.route('/sdg')
-def sdg_page():
-  placeholder_query = ''
-  return render_template('/nl_interface.html',
-                         maps_api_key=current_app.config['MAPS_API_KEY'],
-                         placeholder_query=placeholder_query,
-                         index_type="sdg_ft",
-                         website_hash=os.environ.get("WEBSITE_HASH"))
-
-
 @bp.route('/eval')
 def eval_page():
   if not current_app.config['VERTEX_AI_MODELS']:
