--- static/js/apps/nl_interface/app.tsx
@@ -1,50 +1,0 @@
-/**
- * Copyright 2022 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Main component for NL interface.
- */
-
-import { StoreProvider } from "easy-peasy";
-import React, { useEffect } from "react";
-
-import { store, useStoreActions } from "./app_state";
-import { QueryContext } from "./query_context";
-import { Sidebar } from "./sidebar";
-
-/**
- * Application container
- */
-export function App(): JSX.Element {
-  return (
-    <StoreProvider store={store}>
-      <InitializeApp />
-      <Sidebar />
-      <QueryContext />
-    </StoreProvider>
-  );
-}
-
-/**
- * Dummy component. Initializes application state.
- */
-const InitializeApp: React.FC = () => {
-  const initializeAppState = useStoreActions((a) => a.initializeAppState);
-  useEffect(() => {
-    initializeAppState();
-  }, [initializeAppState]);
-  return null;
-};
