+++ static/js/utils/explore_utils.ts
