+++ static/js/utils/tile_utils.tsx
@@ -23,7 +23,7 @@ import * as d3 from "d3";
 import _ from "lodash";
 import React from "react";
 
-import { NL_SOURCE_REPLACEMENTS } from "../constants/app/nl_interface_constants";
+import { NL_SOURCE_REPLACEMENTS } from "../constants/app/explore_constants";
 import { SELF_PLACE_DCID_PLACEHOLDER } from "../constants/subject_page_constants";
 import { CSV_FIELD_DELIMITER } from "../constants/tile_constants";
 import {
@@ -38,7 +38,7 @@ import { getCappedStatVarDate, urlToDisplayText } from "../shared/util";
 import { getMatchingObservation } from "../tools/shared_util";
 import { EventTypeSpec, TileConfig } from "../types/subject_page_proto_types";
 import { stringifyFn } from "./axios";
-import { isNlInterface } from "./nl_interface_utils";
+import { isNlInterface } from "./explore_utils";
 import { getUnit } from "./stat_metadata_utils";
 
 const DEFAULT_PC_SCALING = 100;
